Separate references into its own service
