+++ app/backend/src/tests/test_fixtures.py
@@ -23,6 +23,7 @@ from couchers.servicers.groups import Groups
 from couchers.servicers.jail import Jail
 from couchers.servicers.media import Media, get_media_auth_interceptor
 from couchers.servicers.pages import Pages
+from couchers.servicers.references import References
 from couchers.servicers.requests import Requests
 from couchers.servicers.search import Search
 from couchers.utils import create_coordinate
@@ -38,6 +39,7 @@ from pb import (
     jail_pb2_grpc,
     media_pb2_grpc,
     pages_pb2_grpc,
+    references_pb2_grpc,
     requests_pb2_grpc,
     search_pb2_grpc,
 )
@@ -376,6 +378,16 @@ def search_session(token):
     channel = fake_channel(token)
     search_pb2_grpc.add_SearchServicer_to_server(Search(), channel)
     yield search_pb2_grpc.SearchStub(channel)
+
+
+@contextmanager
+def references_session(token):
+    """
+    Create a References API for testing, uses the token for auth
+    """
+    channel = fake_channel(token)
+    references_pb2_grpc.add_ReferencesServicer_to_server(References(), channel)
+    yield references_pb2_grpc.ReferencesStub(channel)
 
 
 @contextmanager
